refactor: 필요 없는 class 정리
