+++ src/main/java/toolc/yourlist/play/domain/ReadThumbnail.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.play.infra;
+package toolc.yourlist.play.domain;
 
 interface ReadThumbnail {
   String find(Long playlistId);
