--- src/main/java/toolc/yourlist/play/infra/Play.java
@@ -1,7 +1,0 @@
-package toolc.yourlist.play.infra;
-
-import java.util.List;
-
-public interface Play {
-  List<PlayEntity> readWhatBelongsTo(Long playlistId);
-}
