+++ src/main/java/toolc/yourlist/play/infra/SpringDataJPAPlayRepository.java
@@ -1,6 +1,7 @@
 package toolc.yourlist.play.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import toolc.yourlist.play.domain.JpaPlayRepository;
 
 public interface SpringDataJPAPlayRepository extends JpaRepository<PlayEntity, Long>, JpaPlayRepository {
 }
